refactor teleop node and start on gripper integration
